+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [Fix] Comments added in SQL query to be stripped before saved as snippet (#886)
 * [Fix] Fixed bug passing :NUMBER while string slicing in query (#901)
 * [Fix] Disabled CTE generation when snippets are detected in a non-SELECT type query. (#651, #652)
+* [Fix] Fix empty result in certain duckdb `SELECT` and `SUMMARIZE` queries with leading comments (#892)
 
 ## 0.10.2 (2023-09-22)